captionmg/requirements.txt
Flask
torch
transformers
Pillow
opencv-python
gTTS